pyproject.toml
[project]
name = "dte-triage-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.4.1",
    "pytest-httpserver>=1.1.3",
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "ruff>=0.12.10",
    "streamlit>=1.48.1",
    "watchdog>=6.0.0",
]
